requirements.txt
altair
numpy
pandas
pydeck
streamlit
openai == 1.6.1
openpyxl == 3.1.2
farm-haystack